Improve planet holder
