+++ frontend/src/javascripts/components/atoms/planet.js
@@ -1,12 +1,8 @@
 import React from 'react'
 import { PlanetImgs } from '../../constants/images'
 
-const Planet = ({ assignmentInfo, onClick }) => (
-  <img
-    src={PlanetImgs[assignmentInfo.planet_type]}
-    className="planet"
-    onClick={onClick}
-  />
+const Planet = ({ className, planetType, onClick }) => (
+  <img className={className} src={PlanetImgs[planetType]} onClick={onClick} />
 )
 
 export default Planet
